optimized Code
