+++ src/features/profile/Orders.css
@@ -11,7 +11,6 @@
   font-family: sans-serif; /* Basic font */
 }
 
-
 .nav-list {
   list-style: none;
   padding: 0;
