+++ src/features/profile/Payments.css
@@ -10,15 +10,25 @@
   padding: 20px; /* Padding around the entire addresses section */
   width: 100%; /* Full width of the parent */
   max-width: 2000px; /* Optional: constrain width */
-  margin-left: 0;
   font-family: sans-serif; /* Basic font */
 }
 
+.account-navigation-container {
+  width: 350px;
+  min-height: 100vh; /* Full viewport height */
+  background-color: #f8f8f8;
+  border-right: 1px solid #eee;
+  padding: 20px;
+  left: 0;
+  top: 0;
+  z-index: 1000;
+  overflow-y: auto;
+}
 .payment-methods-content {
   flex-direction: column;
   width: 100%;
 }
-.payment-methods-content h2 {
+.payment-methods-content page-title {
   font-size: 2.5rem;
   font-weight: 700;
   text-align: center;
@@ -80,13 +90,13 @@
   flex: 1;
 }
 
-.payment-option-details h4 {
+.payment-option-details payment-mode {
   margin: 0 0 0.25rem 0;
   font-size: 1rem;
   font-weight: 500;
 }
 
-.payment-option-details p {
+.payment-option-details payment-caption{
   margin: 0;
   font-size: 0.875rem;
   color: #666;
