+++ src/features/profile/ProfilePage.css
@@ -1,0 +1,241 @@
+/* ProfilePage.css */
+/* ProfilePage.css */
+.profile-page-container {
+  display: flex;
+  max-width: 100%; /* Full width */
+  margin: 0;
+  padding: 20px;
+  gap: 0;
+}
+
+.account-navigation-container {
+  width: 300px;
+  min-height: 100vh; /* Full viewport height */
+  background-color: #f8f8f8;
+  border-right: 1px solid #eee;
+  padding: 20px;
+  left: 0;
+  top: 0;
+  z-index: 1000;
+  overflow-y: auto;
+}
+
+.account-navigation-container h3 {
+  margin-top: 0;
+  margin-bottom: 15px;
+  color: #000000;
+  font-size: 1.6rem;
+}
+
+.account-navigation-container ul {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.account-navigation-container ul li {
+  margin-bottom: 5px;
+}
+
+.account-navigation-container ul li a {
+  text-decoration: none;
+  color: #333;
+  display: block;
+  padding: 10px 15px;
+  transition: background-color 0.3s ease, color 0.3s ease;
+  border-radius: 4px;
+}
+
+.account-navigation-container ul li a:hover {
+  background-color: #f0f0f0;
+}
+
+.account-navigation-container ul li a.active {
+  font-weight: bold;
+  color: #007bff;
+  background-color: #e6f2ff;
+}
+
+.profile-card {
+  flex-grow: 1;
+  margin-left: 250px; /* Match the width of navigation container */
+  padding: 24px;
+  max-width: calc(100% - 250px);
+}
+
+/* Responsive adjustments */
+@media (max-width: 768px) {
+  .account-navigation-container {
+    width: 100%;
+    position: static;
+    border-bottom: 1px solid #eee;
+    margin-bottom: 20px;
+  }
+
+  .profile-card {
+    margin-left: 0;
+    max-width: 100%;
+  }
+
+  .profile-page-container {
+    flex-direction: column;
+  }
+}
+
+.account-navigation-container h3 {
+  margin-top: 0;
+  margin-bottom: 15px;
+  color: #000000;
+  font-size: 1.6rem;
+}
+
+.account-navigation-container ul {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.account-navigation-container ul li {
+  padding: 10px 15px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+  font-size: 1.2rem;
+}
+
+.account-navigation-container ul li:hover {
+  background-color: #eee;
+}
+
+.account-navigation-container ul li a {
+  text-decoration: none;
+  color: #333;
+  display: block;
+  padding: 10px 15px;
+  transition: background-color 0.3s ease, color 0.3s ease;
+  border-radius: 4px;
+}
+
+.account-navigation-container ul li a:hover {
+  background-color: #f0f0f0;
+}
+
+.account-navigation-container ul li a.active {
+  font-weight: bold;
+  color: #007bff;
+  background-color: #e6f2ff;
+}
+
+.profile-card {
+  flex-grow: 1;
+  border-radius: 8px;
+  padding: 24px;
+  width: 100%;
+  max-width: 800px;
+}
+
+.profile-card h2 {
+  font-size: 2.5rem;
+  font-weight: 700;
+  text-align: center;
+  color: #000000;
+  margin-bottom: 1rem;
+}
+
+.profile-info {
+  display: flex;
+  flex-direction: column;
+  background-color: #f5f5f5;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  padding: 40px;
+  gap: 15px;
+}
+
+.info-item {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 0;
+  border-bottom: 1px solid #e0e0e0;
+}
+
+.label {
+  font-weight: bold;
+  color: #555;
+  width: 180px;
+}
+
+.value {
+  flex-grow: 1;
+  overflow-wrap: break-word;
+  text-align: left;
+  padding-left: 20px;
+}
+
+.edit-button-container {
+  text-align: right;
+  margin-bottom: 15px;
+}
+
+.edit-button {
+  background-color: #000000;
+  color: white;
+  border: none;
+  padding: 10px 20px;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.3s;
+  font-weight: 500;
+}
+
+.edit-button:hover {
+  background-color: #333;
+}
+
+.add-button {
+  background-color: #000000;
+  color: white;
+  border: none;
+  padding: 8px 15px;
+  border-radius: 4px;
+  cursor: pointer;
+  font-weight: 500;
+}
+
+.add-button:hover {
+  background-color: #333;
+}
+
+.mobile-info {
+  flex-grow: 1;
+}
+
+.security-note {
+  font-size: 0.8rem;
+  color: #666;
+  margin-top: 5px;
+}
+
+.warning-icon {
+  margin-right: 5px;
+}
+
+.error-message {
+  color: #d9534f;
+  margin-top: 20px;
+  padding: 10px;
+  background-color: #f8d7da;
+  border-radius: 4px;
+}
+
+/* Responsive adjustments */
+@media (max-width: 768px) {
+  .profile-page-container {
+    flex-direction: column;
+  }
+
+  .account-navigation-container {
+    width: 100%;
+    border-bottom: 1px solid #eee;
+    margin-bottom: 20px;
+  }
+}
