+++ src/features/profile/ProfilePage.jsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
-import styles from './ProfilePage.module.css';
-import './accountpage.css';
-import { NavLink, useNavigate } from 'react-router-dom';
-import AccountNavigation from './AccountNavigation'; // Adjust the import path as necessary
+import './ProfilePage.css';
+import axios from 'axios';
+import { useNavigate } from 'react-router-dom';
+import AccountNavigation from './AccountNavigation';
 import withLayout from '../../layouts/HOC/withLayout';
 
 const ProfilePageContent = () => {
@@ -12,6 +12,7 @@ const ProfilePageContent = () => {
     mobileNumber: null,
     username: 'Loading...',
   });
+
   const [error, setError] = useState(null);
   const navigate = useNavigate();
 
@@ -26,7 +27,7 @@ const ProfilePageContent = () => {
     setProfile(defaultProfile);
 
     // Fetch profile data from API
-    fetch('/api/profile')
+    axios('/api/profile')
       .then(response => {
         if (!response.ok) {
           throw new Error(`Failed to fetch profile: ${response.status}`);
@@ -37,7 +38,6 @@ const ProfilePageContent = () => {
         setProfile(data);
       })
       .catch(err => {
-        console.error('Error fetching profile:', err);
         setError('Failed to load profile information.');
       });
   }, []);
@@ -51,53 +51,57 @@ const ProfilePageContent = () => {
   };
 
   return (
-    <div className={styles.profileCard}>
-      <div style={{ textAlign: 'right', marginBottom: '15px' }}>
-        <button id="editProfileButton" className={styles.editButton} onClick={handleEditProfileClick}>
+    <div className="profile-card">
+      <div className="edit-button-container">
+        <button 
+          id="editProfileButton" 
+          className="edit-button" 
+          onClick={handleEditProfileClick}
+        >
           Edit Profile
         </button>
       </div>
       <h2>Your Profile</h2>
-      <div className={styles.profileInfo}>
-        <div className={styles.infoItem}>
-          <span className={styles.label}>Name</span>
-          <span className={styles.value}>{profile.fullName || 'N/A'}</span>
+      <div className="profile-info">
+        <div className="info-item">
+          <span className="label">Name</span>
+          <span className="value">{profile.fullName || 'N/A'}</span>
         </div>
-        <div className={styles.infoItem}>
-          <span className={styles.label}>Email</span>
-          <span className={styles.value}>{profile.email || 'N/A'}</span>
+        <div className="info-item">
+          <span className="label">Email</span>
+          <span className="value">{profile.email || 'N/A'}</span>
         </div>
-        <div className={styles.infoItem}>
-          <span className={styles.label}>Mobile number</span>
+        <div className="info-item">
+          <span className="label">Mobile number</span>
           {profile.mobileNumber ? (
-            <div className={styles.mobileInfo}>
-              <span className={styles.value}>{profile.mobileNumber}</span>
-              <p className={styles.securityNote}>
-                <span style={{ color: 'orange', fontSize: '1.2em' }}>⚠️</span>
+            <div className="mobile-info">
+              <span className="value">{profile.mobileNumber}</span>
+              <p className="security-note">
+                <span className="warning-icon">⚠️</span>
                 For stronger account security, add your mobile number. If there's an
                 unusual sign-in, we'll text you and verify that it's really you.
               </p>
             </div>
           ) : (
-            <button className={styles.addButton} onClick={handleAddMobile}>Add</button>
+            <button className="add-button" onClick={handleAddMobile}>Add</button>
           )}
         </div>
-        <div className={styles.infoItem}>
-          <span className={styles.label}>Username</span>
-          <span className={styles.value}>{profile.username || 'N/A'}</span>
+        <div className="info-item">
+          <span className="label">Username</span>
+          <span className="value">{profile.username || 'N/A'}</span>
         </div>
-        <div className={styles.infoItem}>
-          <span className={styles.label}>Password</span>
-          <span className={styles.value}>********</span>
+        <div className="info-item">
+          <span className="label">Password</span>
+          <span className="value">********</span>
         </div>
       </div>
-      {error && <p className={styles.errorMessage}>{error}</p>}
+      {error && <p className="error-message">{error}</p>}
     </div>
   );
 };
 
 const ProfilePage = () => (
-  <div className={styles.profilePageContainer}>
+  <div className="profile-page-container">
     <AccountNavigation />
     <ProfilePageContent />
   </div>
