+++ src/routes/AppRouter.jsx
@@ -4,12 +4,12 @@ import LandingPage from '../features/home/LandingPage.jsx';
 import Product from '../features/products/Products.jsx';
 import Login from '../features/auth/login.jsx';
 import ProfilePage from '../features/profile/ProfilePage.jsx';
-import AccountPage from '../features/profile/accountpage.jsx'; // Importing AccountPage
-import EditProfile from '../features/profile/edit-profile.jsx';
-import ManageAddressesPage from '../features/profile/manageaddress.jsx';
-import OrderManager from '../features/profile/orders.jsx';
-import PaymentMethods from '../features/profile/payments.jsx';
-import Help from '../features/profile/help.jsx';
+import AccountPage from '../features/profile/AccountPage.jsx'; // Importing AccountPage
+import EditProfile from '../features/profile/Edit-Profile.jsx';
+import ManageAddressesPage from '../features/profile/ManageAddress.jsx';
+import OrderManager from '../features/profile/Orders.jsx';
+import PaymentMethods from '../features/profile/Payments.jsx';
+import Help from '../features/profile/Help.jsx';
 import CartPage from '../features/cart/CartPage.jsx';
 import ProductPage from '../features/products/ProductPage.jsx';
